firmware_TouchMTB: take MtbSanityValidator as a prerequisite
If the MtbSanityValidator fails in a test, it indicates the existence
of some critical problems. The rest validators of the test would not
be executed because some unexpected problems or results might be
incurred.

BUG=chrome-os-partner:30965
TEST=Record gesture files on a chromebook machine.
$ cd /usr/local/autotest/tests/firmware_TouchMTB/
$ DISPLAY=:0 python main.py -m complete

Before pressing ENTER for the 2nd command above, place 2 fingers on
the touchpad. Then press ENTER. It would only show the test results
about MTB Sanity on the right side of the gesture capturing window.
There are no test results about the other validators.

. . .
  MtbSanityValidator
    # fingers: 2
    slot 0:
      MT_TOUCH_MAJOR=   254
      MT_TRACKING_ID=     4
      MT_PRESSURE=    52
      MT_POSITION_X=   573
      MT_POSITION_Y= 299
    slot 1:
      MT_TOUCH_MAJOR= 254
      MT_TRACKING_ID= 5
      MT_PRESSURE= 57
      MT_POSITION_X= 1384
      MT_POSITION_Y= 421
    # MTB # errors: # 473
    FINGER_LEAVING_BEFORE_ARRIVING: 2
    ASSIGN_ATTRIBUTES_BEFORE_ARRIVING: 471
    criteria_str: == 0
  score: 0.000000
.  .  .

Change-Id: Ic89fb78eb36ef911521adf094eafb9857e218e32
Reviewed-on: https://chromium-review.googlesource.com/210791
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>