Compare images locally before uploading to biopic.
For our video tests we are currently using biopic to compare screenshots
of the video.

In some cases we end up uploading alot of images to biopic,
this overloads the server, especially considering that we don't need
to upload images which have passed the test.
We could use biopic verification for test images that have failed the test.

One solution to this is to use local comparison first and if that test fails,
then upload to biopic so we can look at it later.

In this CL:
Introduce a local RGB comparer.
Introduce a double comparer comparer. (See what I did there.)
Unify the comparer interface (duck type).
Use the new stuff in the video glitch detection test.

BUG=chromium:393684
TEST=Ran video tests locally. Also locally tested the comparers.

Change-Id: I144d5196b0c259453fe3a254c339da776acad91f
Reviewed-on: https://chromium-review.googlesource.com/210400
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>