Clean up cros_ui.restart() API
cros_ui.restart() took a callable to allow a custom mechanism
for restarting the UI. The only thing using this was passing
in a pyauto method, which was deprecated ages ago. Remove this
but allow instead for the caller to specify that it's OK
for the UI to be down already when calling cros_ui.restart().

BUG=None
TEST=stop ui and then run login_OwnershipRetaken

Change-Id: I78286a5edc777f6a96e2b48e13d386c3a9fd6d0b
Reviewed-on: https://chromium-review.googlesource.com/210054
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>