autoupdate_EndToEndTest: Collect and report performance data.
While performing an update, run a script on the DUT to measure how
many resources update_engine and its helpers are using. For now we
only record the peak memory usage but the performance monitoring
script can easily be extended to record other things.

For now, we report all samples under the same name which means that
multiple samples from the same source version are reported (one for
the full update and one for each delta). In the future we might want
to report this separately.

BUG=chromium:385305
TEST=Ran the autoupdate_EndToEndTest on Moblab w/o any problems.

Change-Id: I866068c83fb755b283e4b41e1828284e6d8cbcca
Reviewed-on: https://chromium-review.googlesource.com/205434
Tested-by: David Zeuthen <zeuthen@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>