Don't make any servo XMLPRC calls during Servo contructor.
In order to test servo, it can be necessary to create a servo object
for a servo host where servod isn't working.  This changes the Servo
constructor not to fail in that case, allowing the callers to have
control over when and how to test for a working servod.

BUG=chromium:388458
TEST=run various test cases on a local instance

Change-Id: I8e5c51be2287ec4b93a6519144540c90846b88d4
Reviewed-on: https://chromium-review.googlesource.com/211875
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>