[autotest] Add script that creates alias to an index in esdb
This script will create an alias for an index in elasticsearch db.
For example, if we have an index, cautotest_backup and we want to
refer to it by the name cautotest, then this script will be handy.
This script will mainly be used after creating a new index.

BUG=None
TEST=./es_create_alias.sh -s 172.25.61.45 -p 9200 -i test_index -a test_index_alias

Change-Id: I6c39f0cd85e7079a42dcc305d1168ea781ccb75c
Reviewed-on: https://chromium-review.googlesource.com/212112
Tested-by: Michael Liang <michaelliang@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Michael Liang <michaelliang@chromium.org>