Crop images to desired dimensions before comparison.
Images contain a text label showing the time at which the video was sampled.
This was left to assist later analysis. However, it takes a dependency on
text rendering that changes and breaks the tests.

We are cropping the images to exclude this label when doing a local RGB
comparison. If the comparison is successful then we return, if it is not
then we upload to bp as before.

** Together with this change, force images collected to be RGB.

BUG=chromium:393684
TEST=Ran local test. Should get more coverage from lab.

Change-Id: Ia95679066fc44a3970ff014cee1c113b032a1145
Reviewed-on: https://chromium-review.googlesource.com/212190
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>