[autotest] rpm frontend server retrieves rpm/outlet/hydra from afe
This is the 2nd CL to make rpm infra read rpm/outlet/hydra info from
afe.

The major change of this cl is to add a new method
    _get_powerunit_info(...)
For chromeos duts, it first looks up in its local lru cache for
rpm/outlet/hydra information for the dut. If not found in
local cache, retrieve the information from afe.

For servos, it loads the poe/outlet/hydra info from a local file.

The method returns an instance of PowerUnitInfo instance, which
will be passed to rpm_dispatcher.

CQ-DEPEND=CL:212349
BUG=chromium:392548
TEST=unittest.
Integration test with other cls in this series, set up
rpm server locally and change host states.

Change-Id: I63326f8b58083ebf807105d94c919bf5fd5f7cea
Reviewed-on: https://chromium-review.googlesource.com/212346
Reviewed-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>