Adding the changes to the two Netgear routers.
This is a copy of https://chromium-review.googlesource.com/#/c/212041.

NetgearR6200 was not loading the page completely. Now we make sure that its
fully loaded before we proceed.

Netgear4300 still has issues, because it does load the page now but will change
the frame immediately after it finishes loading the wireless admin page.
This needs a better resolution. For now I have made sure that we wait till the
page loads and have used JS to top the page from moving to a different frame
after it finishes loading, but this is not working currently from the webdriver.

Netger2000 didn't show any errors when I ran the unittest against it. The error
that you have seen might have been a one off issue.

Belkin F9K1102 was not allowing open security in N mode, hence we are removing
Wireless N mode from the list of supported modes.

BUG=chromium:388911,381480,388967
TEST=unittest

Change-Id: I4f9f78245409348b1e43f136087edecdb44a912a
Reviewed-on: https://chromium-review.googlesource.com/212470
Reviewed-by: Deepak Gopal <deepakg@chromium.org>
Commit-Queue: Deepak Gopal <deepakg@chromium.org>
Tested-by: Deepak Gopal <deepakg@chromium.org>