Replacing sync and sleep with a blocking sync call.
Currently we call sync and then sleep for expected time required. This
seems to be problematic in situations when the expected sleep time is
too short.

Instead we are faking a blocking sync by calling sync twice. The first
time its called, it returns before the sync is complete, but the
subsequent call should wait for the first one to complete.

Also, removed the sync faft config value since its not used anymore.

BUG=chromium:239034,chromium:31300
TEST=ran tests at desk

Change-Id: I0b122f80a62fd5a1442b7b3c8e205e1a78bf4a59
Reviewed-on: https://chromium-review.googlesource.com/212492
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>