firmware_TouchMTB: fix a bug when touch_device is None
It is handy to be able to conduct unittests in chroot especially
when a chromebook is not in hand. This patch fixed a bug caused
by the fact that there might be no touch device on a desktop host.

BUG=chrome-os-partner:30965
TEST=None. Unittests should all pass when executed in chroot.
(cr) $ cd .../autotest/files/client/site_tests/firmware_TouchMTB
(cr) $ python tests/run_all_unittests.py

Change-Id: Ie8a0db3857fc94e7ace7cec489d85257a3467197
Reviewed-on: https://chromium-review.googlesource.com/212851
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>