Don't fail test for pixels that are practically black.
In an RGB image a pixel is purely black if its RGB values are (0, 0, 0). Ideally, the difference image (golden - test run) should be purely black. This is not always the case. Sometimes the diff image has pixels that are mostly black. In those cases the two images are different but the difference is indistinguishable by human eye.

We don't need to fail these comparison.

Introduce a threshold parameter which is the maximum value between (0, 256) that a color component is allowed to have. Any color component that is above the threshold causes that pixel to be counted as a 'fail' pixel.

As usual, this parameter is configurable on a text file. The number was chosen as (30, 30, 30) is practically a black pixel.(~88% black)

BUG=chromium:393684
TEST=Ran test on a daisy (webm, 480p) which is currently being failed in the lab. The test passed as all diff pixels are below threshold.

Change-Id: Id6c0656ba80920a605b9e89cee3d053c27bf9eca
Reviewed-on: https://chromium-review.googlesource.com/212710
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>