Added a new validator that discards the initial seconds of data before validation.
The new validator can wrap around any generic validator, and the number
of seconds to discard can be specified.

The validator simply strips the first second and prepends
the end state of the discarded second onto the stripped data. This
throws away all events that happened within the first second, but keeps
all events that persist after the first second.

The noise tests also now have a one second pause at the beginning of
each test to allow the touchpad to calibrate against the noise.

BUG=chromium:404265
TEST=Manually and with robot

Change-Id: I6d6a0c40f4cb106185c7cc084fca3b459dd49c3c
Signed-off-by: Mindy Huang <mindyh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/212763
Reviewed-by: Charles Mooney <charliemooney@chromium.org>