firmware_TouchMTB: a temporary hack of installing pygtk for x86_64
The packages pygtk and gtk+ have been removed. As a temporary
work-around, install the related scripts and binaries from a tarball.
A longer-term solution would be to move the GUI to Chrome HTML + JS.

BUG=chromium:404570
TEST=Using the latest test image. The touch firmware test should
execute normally with a gtk window to guide the user to perform
various gestures.

Change-Id: I8d8f670c3ae2b4aa07594b2a151277b2f32bdfbc
Reviewed-on: https://chromium-review.googlesource.com/213080
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>