[autotest] Check servod process's status in verify_software
We've seen servo verification took very long time when servod is actually not
running. This adds a lot of overhead to the test. The timeout can be as long
as 2 minutes and servo verification is done 3 times in each test.

This CL add a check of actual servod process in servo_host.verify_software.
If servod is not running, servo host object will fail to be verified without
waiting for actual servo call to timeout in servo.initialize_dut, e.g.,
self.set('dut_hub_pwren', 'on')

Also add a timeout around servo.initialize_dut for 30 seconds.

BUG=chromium:404209
TEST=local servo setup and test_that

Change-Id: I46d2d10cd188db57c73b749c688d3d788a6df305
Reviewed-on: https://chromium-review.googlesource.com/213671
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>