hardware_RamFio: Don't raise error when free memory is not enough.
We determine available memory size using more conservative method.
For machine with 2GB Ram. Sometimes the remaining memory is not enough
for 1GB + 20% overhead require hardware_RamFio. This caused test
glake in the lab.

This CL will just logging if there is not enough memory instead and
use what is available to test.

BUG=chromium:402530
TEST=run using test_that, hardware_RamFio don't raise error.

Change-Id: I749eeadeb3966a73ab370d615b74e27a13524f72
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/213563
Reviewed-by: Jay Kim <yongjaek@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>