Revert "[autotest] Add shard tables and models to Autotest"
Afe jobs has 7 million rows and adding a column might take  upwards of 20 minutes. I would like to avoid the downtime
during the day, and we need a push asap for a servo fix 
that's affecting cq. Since this isn't critical, I'm moving
it out of the way till we have space for the migrate.

This reverts commit 962e3413dce0a6ea2fb0bed87323655f1641e57c.

Change-Id: I053c59cdab314aa7bab5db37b3a4d7cada1f1c7d
Reviewed-on: https://chromium-review.googlesource.com/213791
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>