Added a unit test for DiscardInitialSecondsValidator and fixed various bugs with the validator.
Added in a check to prevent crashing if the tid_packet returned from the state
machine was None.

Also changed a parameter with the state machine so that it returns a value even
if all the events in the packet have not been received. This prevents the bug where
a finger is not included in the final state because a coordinate or pressure event
had not been received.

BUG=chromium:407196
TEST=Manually

Change-Id: I16dbe383653c5e3c066617835e5930992fccd106
Signed-off-by: Mindy Huang <mindyh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214014
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>