[autotest] Moblab Host Verify/Check_device/Repair Proceedures.
This is the initial version of Moblab's Verify and Check Device.

Verify checks that all the moblab services are running and runs
verify on Moblab's DUTs. Verify fails if this Moblab has no DUTs
in the Ready state.

Repair backs up the dhcpd lease file.

check_device for Moblab requires a reboot and creation of the
autotest client directory. If the dhcpd lease file was backed up,
it is now restored.

find_and_add_duts switches from broadcasting on the network to
directly pinging DUTs listed in the dhcpd lease file.

Updated the stable cros version to 6170.0.0 from 6167.0.0 as
6167.0.0 for Moblab has a bug critical to Moblab testing.

BUG=chromium:388462
TEST=Verify/Repair run successful. moblab_SmokeSuite test. Ran
bvt-inline without reimage.

Change-Id: Idb351d67eda70a95960c1d9f2e08ca586aad73bf
Reviewed-on: https://chromium-review.googlesource.com/214359
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>