[autotest] suite_scheduler: Don't die on a TimeoutException.
If we happen to not be able to contact the AFE, just conservatively
assume that we shouldn't run that request.  This way, if we happen to
time out while scheduling nightly suites, we don't just crash and give
up on scheduling all of them.

BUG=chromium:407701
DEPLOY=suite_scheduler
TEST=unit

Change-Id: I29884be237cc8697f614352295e5b7891e64de7f
Reviewed-on: https://chromium-review.googlesource.com/214322
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>