Add verification to ui_SystemTray tests.
Initially, we had the 'test' only collect screenshots of the
system tray with the intent to watch how the tray varies and
assess the general viability of image comparison based tests
in this area.

After watching the screenshots for a number of builds, it makes
sense to start with just the language indicator portion of the
tray.

This change collects this particular portion of the system tray,
pulls down a golden version of it and runs the comparison.

Comparisons are done locally first with a simple pixel by pixel
array walk. Failures are uploaded to bp for later viewing.

BUG=chromium:374981
TEST=Ran tests against a local link.
TEST=Ran video glitch detection tests also.

Change-Id: Ic5316d4d17a125b878c3c22b86568c554cf41186
Reviewed-on: https://chromium-review.googlesource.com/214380
Reviewed-by: Tracy Turchetto <tturchetto@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>