Adding a check that dut is down during full power cycle.
Currently it presses the power button, wait and presses again. We don't
confirm that the DUT is actually down. This might be a problem if there
is an infrastructure issues where the system doesn't actually turn off.

Keeping the sleep, since there might be a need that we wait for a
minimum amount of time for the device to shutdown fully.

TEST=ran at desk against winky
BUG=chromium:408297
DEPLOY=test_importer

Change-Id: I732aaff5aaf1024f0de10fb90d3e8a107e13e341
Reviewed-on: https://chromium-review.googlesource.com/214593
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>