[autotest] Make global scheduler not execute hqes of shards
The scheduler today just executes all hqes that are in the database.
With shards this behaviour is not wanted, it should only execute
the hqes that are not meant to be executed by a shard.

This changes adds a condition, so the scheduler won't pick up hqes
with meta_hosts, that have an entry in afe_shards_labels.

TEST=Ran suites + Ran dummy, ensured only hostless job is executed.
DEPLOY=scheduler,host_scheduler
CQ-DEPEND=CL:211214

Change-Id: I85b14acb24bb24e9a14d5db2d42c6611fd15fc02
Reviewed-on: https://chromium-review.googlesource.com/214656
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>