video_VDAPerf - Update how we count the dropped frames.
In the original version, vda_unittest won't log the delivery time (actually
the rendering time) for dropped frames. So, we can simply count the number of
logged entries and know the frame drop rate. But we change the behavior
recently. We need to analyze the delivery time and see if it is too late for
rendering.

BUG=chromium:411123
TEST=Run the test on CrOS device

Change-Id: I67e93c7b687f9e0f563a8b9977e04c2c7d47d7df
Reviewed-on: https://chromium-review.googlesource.com/215442
Reviewed-by: Owen Lin <owenlin@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>