Adding duplicate stress suite entries for 4 more pools.
We want to run all the tests in stress suite on 5 machines of the same
board type. Currently there is no way to specify this. Faking it by
creating duplicate entries for the stress suite, pointing to different
pool labels which we will add to the duplicate machines.

TEST=None
BUG=chromium:396322

Change-Id: I0e6b08e72cad0ea5946c6cbf7708d9b221aa9b75
Reviewed-on: https://chromium-review.googlesource.com/214835
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>