[autotest] Add a --production argument to stop scheduler from starting with database set to remote server
This is the first CL to add the option. Once this CL is merged and pushed to
prod. A second CL will update the upstart job in puppet to use the option.
Once puppet change is populated to prod, a third CL will remove the override
of production argument and make it effective in prod.

BUG=chromium:409091
TEST=local setup, change host in AUTOTEST_WEB to various settings and verify
the option works.
DEPLOY=scheduler,host-scheduler

Change-Id: I0b3b961db3ae7a95b887f5ae32b73c3a33b6e82c
Reviewed-on: https://chromium-review.googlesource.com/215672
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>