USB multiplexer controller class.
USBMuxController is used to control the individual ports on a 8 port USB
hub/switch.This class is responsible for enabling all the GPIO pins on
the needed to control the 8 port USB switch/hub. In order to use this
USB mux controller you need custom hardware setup which connects to the
beaglebone and drives the 8 port relay switch to turn the individual
ports on the USB hub 'on'/'off'.

BUG=chromium:409275
TEST=Tested against a local DUT

Change-Id: If034a63b5b19988e4c25d3f1f7a50fff0eee9303
Reviewed-on: https://chromium-review.googlesource.com/215692
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: harpreet Grewal <harpreet@chromium.org>
Commit-Queue: harpreet Grewal <harpreet@chromium.org>
Tested-by: harpreet Grewal <harpreet@chromium.org>