Make sure Servo is initialized before checking the USB stick state.
One of the verification steps for Servo is to confirm that the
USB stick isn't wedged as for crbug.com/225932.  The check
pre-supposed standard initialization, which unplugs the USB stick,
but the initialization didn't happen until after the check.  This
made it possible for units to fail incorrectly in some cases.

This change re-orders the code to insure initialization prior to
the USB check.

BUG=chromium:409292
TEST=trigger repair in local instance with a plugged-in, corrupted USB stick

Change-Id: I8c8099742cfc32fcd1f7f3f3669d801924011dac
Reviewed-on: https://chromium-review.googlesource.com/215698
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>