[autotest] Add shard_id to hosts
To assign hosts to shards and avoid reassigning them again and again
all the time, somehow it must be stored if a host was already synced
to a shard.

This changelist adds a column in afe_hosts to reference the shard
it has been synced to.

TEST=Ran suites and manually tried db migrations and running dummy.
DEPLOY=scheduler,apache,host-scheduler,migrate

Change-Id: Ide181a221f363fe96130611da7732baec9680a4f
Reviewed-on: https://chromium-review.googlesource.com/215997
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>