[autotest] Add serialization functionality including dependencies
In order to send hosts and jobs to shards it's necessary to
serialize them including their dependencies.

This changelist adds functionality to find an object's dependencies
and to serialize it into lists and dicts, so it can be converted
to json.

TEST=Ran suites.
DEPLOY=apache

Change-Id: Icc158de1f23c7483ebc22b6ca3482b7ffccdd5c4
Reviewed-on: https://chromium-review.googlesource.com/215998
Tested-by: Jakob Jülich <jakobjuelich@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Jakob Jülich <jakobjuelich@chromium.org>