[autotest] Don't allow label detection to break host creation
This changes CrosHost.get_labels() so that if a label detection
function raises an exception, the failure is logged and dropped.
This means that individual labels may be omitted because of bugs
in the label detection code, but that the host will still be added
with other detectable labels.

BUG=chromium:410620
TEST=run `atest host create` on a power washed device, see exception but no failure

Change-Id: I6a2a02167776077283fa6bc1d5f46824703d3098
Reviewed-on: https://chromium-review.googlesource.com/216361
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>