[autotest] Fix log socket server unittest
The unittest works as:
1. start a process listen to a local socket, log any message received.
2. start 10 processes to write log to that socket.

It used to be that if the number of logs does not exactly match, the unittest
will fail. As it could be flake due to local network issue, one or two log
attempts could fail. This CL relaxes the check to allow the test to pass if
any message was received by log server.

BUG=chromium:408028
TEST=local unittest

Change-Id: I4392b4dba5a2da3812cf1e6756016dcd20053a11
Reviewed-on: https://chromium-review.googlesource.com/216494
Commit-Queue: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>