[autotest] Update the stable build to R39-6231.0.0
This updates the default build used for repair to a build known
not to suffer from the "demo mode bug".  The build has image
artifacts for all boards in the lab.  The build passed bvt-inline
for most boards, failed none, but didn't run for the following
boards:
    mccloud, peach_pi, wolf, zako

BUG=chromium:398271
TEST=None (but see above regarding build status)

Change-Id: I8d210d9b92f75d738977be79c6591bf9593e94d5
Reviewed-on: https://chromium-review.googlesource.com/216495
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mungyung Ryu <mkryu@google.com>