autoupdate_EndToEndTest: Remove '\n' from error message.
When a test fails and a bug is autofiled, only the first line of
the error message is included in the ANCHOR which will duplicate
bugs that failed for different reasons. This patch removes the \n
before the error message from stderr when raising an
OmahaDevserverFailedToStart exception.

BUG=chromium:409357
TEST=None, just the linter.

Change-Id: I0f859008a54a2c1a6e07a1b763314386cb012193
Reviewed-on: https://chromium-review.googlesource.com/216580
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>