autoupdate_EndToEndTest: Log stderr from devserver.py to autotest log.
This patch redirect the stdout and stderr from devserver.py to a file
that is later dumped togheter with the devserver.py log. Normally,
devserver.py should ouput all its log information to a logfile passed
in the command line. Nevertheless, if there's an error or exception
before it gets a chance to redirect the output to the log, the error
messages in stderr and stdout are lost.

BUG=chromium:370302
TEST=TBA

Change-Id: I4694ac286b4f53705d42a15f972dac1fb2d3ae8e
Reviewed-on: https://chromium-review.googlesource.com/216830
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>