[autotest] Backup hostnames nightly.
Take a nightly backup of just hostnames and the pools they belong to,
and upload it to google storage. In case something goes wrong with
the database server this is the quickest way to recover if one
doesn't care about the preservation of past and present jobs.

TEST=Ran script against local database.
BUG=chromium:410618

Change-Id: I7eb72207fab2fff3bf3f56d5bfcf88ec2356e0b4
Reviewed-on: https://chromium-review.googlesource.com/216854
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>