autoupdate_EndToEndTest: Increase devserver launch timeout.
The autoupdate_EndToEndTest launches a private instance of the
devserver.py in the same devserver, listening in a different port.
The test waits for 15 seconds until the new devserver.py writes its
pid and port number to some temp files. Under heavy load, this value
could be too small.

This patch increases the timeout to 30 seconds and improves the log
message description to include the timeout and the devserver load.

BUG=chromium:370302
TEST=python linter.

Change-Id: I934a6eb926143ccaae2b3bae67d1282361defd2b
Reviewed-on: https://chromium-review.googlesource.com/217081
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>