Clean strings before upload to perf dashboard.
platform_MemoryMonitor is currently failing and not reporting perf
values to the dashboard due to bad characters in the string. This
is a common mistake and usually corrected by replacing bad
characters. We use as default '_' but allow a test to override this.

Furthermore, remove unused variables and do some pylint housekeeping.

BUG=None.
TEST=pylint

Change-Id: I833e08724a9b5b8f684117807ad8d25101f7c5c5
Reviewed-on: https://chromium-review.googlesource.com/217139
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>