[autotest] Drop the USB stick check from ServoHost.verify_software().
The check was meant to look out for regressions in an old bug seen
on Ubuntu kernels, chromium:225932.  However, that problem hasn't
been seen in ages, and maintaining code that can _correctly_
find regressions is now causing more problems than it solves.

So, we're deleting the code.

BUG=chrome-os-partner:31799
TEST=None

Change-Id: I3788e45c2f4ee39b9e883f1fc505430ab5051376
Reviewed-on: https://chromium-review.googlesource.com/217499
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>