In servo-stat, protect against long running ssh commands.
It turns out that one way a beaglebone can misbehave is to accept an
ssh connection and then sit on it indefinitely.  The result is that
`ssh root@$SERVO true` waits forever.  This change adds a timeout to
the 'ssh' command, to make sure that can't happen.

BUG=None
TEST=None

Change-Id: If23036148ed9443d370da9899c7fb430e9ce1db2
Reviewed-on: https://chromium-review.googlesource.com/217712
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>