Autotest: new touch test for touchpad/touchscreen wakeup properties.
Test to look at a DUT's touchpad and touchscreen (if any) and check if they
are wakeup sources.  Currently, touchscreens should never be a wake source
and touchpads should always be one, except for glimmer and clapper (in which
case neither should be).  Found crbug.com/413055 while writing this test,
so it's probably good to have.

Skipping boards with Synaptics touchpads, since they don't report this status.

Also adding this test and the other touch_ tests to a new suite for later
scheduling purposes.

TEST=this; ran on multiple device types
BUG=None

Change-Id: Id5347485b083779d7de03eb3913c6bd8c057bcaa
Reviewed-on: https://chromium-review.googlesource.com/217781
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>