[autotest] run_suite reports provision failure as infra failure
Since most provision jobs failure are due to infra failure,
report infra failure exit code when it happens.

This is a temporary workaround for crbug.com/370302 to make
cq be able to work.

Note the risk introduced here is the possiblity of incorrectly
categorizing a bad build as infra failure,
which is undesired for canary builds.

TEST=run_sutie.py -m JOB_ID on existing suites.
BUG=chromium:370302

Change-Id: I8f4f803e28db41419b2e5d3a567a5c52e97afc1f
Reviewed-on: https://chromium-review.googlesource.com/217934
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>