Force scale factor to value used for golden images.
There was a change in the default scale factor that caused video images
captured from hd devices: nyan and peach pi to appear larger than their
corresponding reference images. This caused the tests to fail. We are
fixing that in this CL.

Applying these args on link will also change the underlying resolution.
We therefore don't do that if the dut is a link.

BUG=chromium:404773
TEST=Ran test on peach_pi and nyan_big without the change, confirmed
that they fail locally as they do in the lab.
Made the change and rann on same devices and ensured that the test
passed.
Ran with change on SD device and ensured the tests passed. (The scale
factor change should not affect SD devices.)
Ran on link with the change ensured tests also passed.

Change-Id: I3b3a0a91b3f27a074ad9fb5b3fdb721473190d40
Reviewed-on: https://chromium-review.googlesource.com/217996
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>